Rename `replicon_channels` module into `channel`
Shorter and fits better since it contains all channel-related stuff,
including the `Channel` enum.
